+++ PandoraWeatherProject/PandoraWeatherProject/Controller/CurrentWeatherViewController.swift
@@ -1,0 +1,59 @@
+//
+//  CurrentWeatherViewController.swift
+//  PandoraWeatherProject
+//
+//  Created by Alex Rhodes on 3/1/20.
+//  Copyright © 2020 Alexandra Rhodes. All rights reserved.
+//
+
+import UIKit
+
+class CurrentWeatherViewController: UIViewController {
+
+    @IBOutlet weak var weatherSegmentedControl: UISegmentedControl!
+    @IBOutlet weak var highLabel: UILabel!
+    @IBOutlet weak var lowLabel: UILabel!
+    @IBOutlet weak var windSpeedLabel: UILabel!
+    @IBOutlet weak var cloudPercentageLabel: UILabel!
+    @IBOutlet weak var sunriseLabel: UILabel!
+    @IBOutlet weak var sunsetLabel: UILabel!
+    @IBOutlet weak var weatherIconImageView: UIImageView!
+    @IBOutlet weak var changeLocationButton: UIBarButtonItem!
+    @IBOutlet weak var favoriteButton: UIBarButtonItem!
+    @IBOutlet weak var currentTempLabel: UILabel!
+    @IBOutlet weak var weatherTypeLabel: UILabel!
+    @IBOutlet weak var forecastedWeatherTableView: UITableView!
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
+
+// MARK: TableView Data Source Methods
+
+extension CurrentWeatherViewController: UITableViewDataSource {
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 5
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        return UITableViewCell()
+    }
+    
+    
+}
