--- PandoraWeatherProject/PandoraWeatherProject/View Controllers/SearchViewController.swift
@@ -1,30 +1,0 @@
-//
-//  SearchViewController.swift
-//  PandoraWeatherProject
-//
-//  Created by Alex Rhodes on 3/1/20.
-//  Copyright © 2020 Alexandra Rhodes. All rights reserved.
-//
-
-import UIKit
-
-class SearchViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
